changes average

+++ src/SalesManager.java
@@ -15,9 +15,9 @@ public class SalesManager {
         return max;
     }
 
-    public int min() {
-        int min = sales[0];
-        for (int sale : sales) {
+    public long min() {
+        long min = sales[0];
+        for (long sale : sales) {
             if (sale < min) {
                 min = sale;
             }
@@ -25,11 +25,11 @@ public class SalesManager {
         return min;
     }
 
-    public int average() {
-        int average = 0;
+    public long average() {
+        long average = 0;
         if (sales.length > 0) {
-            int sum = 0;
-            for (int sale : sales) {
+            long sum = 0;
+            for (long sale : sales) {
                 sum += sale;
             }
             average = (sum - max() - min()) / sales.length;
